добавил сущность handling с полями из handling.meta для транспорта скрытие отобржение интерфейса через АПИ плавное скрытие/отображение интерфейса

+++ common/src/main/kotlin/online/fivem/common/events/net/SpeedometerUpdateEvent.kt
@@ -11,7 +11,7 @@ class SpeedometerUpdateEvent(
 	val handbrake: Boolean,
 	val oilLevel: Float,
 	val petrolTankHealth: Double,
-	val turboPressure: Float?,
+	val turboPressure: Int?,
 	val engineRunning: Boolean,
 	val engineOn: Boolean,
 	val engineHealth: Double
